# docs/swagger.yaml
basePath: /api/
definitions:
...
  http.LoginRequest:
    properties:
      email:
        maxLength: 60
        type: string
      password:
        minLength: 6
        type: string
    required:
    - password
    type: object
  http.UpdateUserRequest:
    properties:
      address:
        maxLength: 100
        type: string
      email:
        maxLength: 60
        type: string
      name:
        maxLength: 60
        minLength: 2
        type: string
      password:
        maxLength: 256
        minLength: 6
        type: string
      patronymic:
        maxLength: 60
        minLength: 2
        type: string
      surname:
        maxLength: 60
        minLength: 2
        type: string
    type: object
  httpErrors.RestError:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
...
  models.User:
    properties:
      address:
        maxLength: 100
        type: string
      admin:
        type: boolean
      email:
        maxLength: 60
        type: string
      id:
        type: integer
      name:
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      patronymic:
        type: string
      surname:
        maxLength: 60
        minLength: 2
        type: string
    required:
    - address
    - name
    - password
    - surname
    type: object
...
  models.UserWithToken:
    properties:
      address:
        maxLength: 100
        type: string
      admin:
        type: boolean
      email:
        maxLength: 60
        type: string
      id:
        type: integer
      name:
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      patronymic:
        type: string
      surname:
        maxLength: 60
        minLength: 2
        type: string
      token:
        type: string
    required:
    - address
    - name
    - password
    - surname
    type: object
  utils.Response:
    properties:
      ok:
        type: boolean
    type: object
  utils.UsersQuery:
    properties:
      address:
        type: string
      email:
        type: string
      limit:
        type: integer
      max_id:
        type: integer
      min_id:
        type: integer
      name:
        type: string
      page:
        type: integer
      patronymic:
        type: string
      surname:
        type: string
    type: object
  utils.UsersQueryResponse:
    properties:
      count:
        type: integer
      page:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
host: localhost:80
info:
  contact:
    email: armangokka@gmail.com
    name: API Support
    url: https://t.me/armanokka
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Time tracker REST API
  version: "1.0"
paths:
...
  /users:
    post:
      consumes:
      - application/json
      description: Register user
      parameters:
      - description: new user info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserWithToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Register user
      tags:
      - auth
  /users/:
    get:
      description: Search users
      parameters:
      - description: search query
        in: body
        name: searchUserQuery
        required: true
        schema:
          $ref: '#/definitions/utils.UsersQuery'
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.UsersQueryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Search users
      tags:
      - auth
  /users/{user_id}:
    delete:
      description: Delete user
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Delete user
      tags:
      - auth
    get:
      description: Get user by ID
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Get user by ID
      tags:
      - auth
    patch:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: string
      - description: new user info
        in: body
        name: updateBody
        required: true
        schema:
          $ref: '#/definitions/http.UpdateUserRequest'
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Update user
      tags:
      - auth
  /users/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: email and password json object
        in: body
        name: searchUserQuery
        required: true
        schema:
          $ref: '#/definitions/http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserWithToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Login user
      tags:
      - auth
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"
...